Persist logic moved to model file & App.tsx moved to src folder & yelp prefixes removed
