+++ src/assets/svgs/index.ts
@@ -1,4 +1,4 @@
-import YelpLogo from './yelp.svg';
+import Logo from './logo.svg';
 import ActiveSVG from './active.svg';
 import AutomotiveSVG from './automotives.svg';
 import BarberSVG from './barbers.svg';
@@ -8,7 +8,7 @@ import RestaurantSVG from './restaurants.svg';
 import ShoppingSVG from './shopping.svg';
 
 const SVGS = {
-  YelpLogo: YelpLogo,
+  Logo: Logo,
   Active: ActiveSVG,
   Automotive: AutomotiveSVG,
   Barber: BarberSVG,
